// src/core/constants/app-constants.enum.ts
export enum AppConstantsEnum {
  PORT = 'PORT',
  ENV_DB_SSL = 'DB_SSL',
  DATABASE_URL = 'DATABASE_URL',
  ENV_DB_HOST = 'DB_HOST',
  ENV_DB_NAME = 'DB_NAME',
  ENV_DB_USER = 'DB_USER',
  ENV_DB_PWD = 'DB_PWD',
  ENV_DB_SCHEMA = 'DB_SCHEMA',
  ENV_JWT_SECRET = 'JWT_SECRET',
  JWT_EXPIRES_TIME = 60,
  SMTP_EMAIL = 'SMTP_EMAIL',
  SMTP_PWD = 'SMTP_PWD',
  PATH_API = 'PATH_API',
  PATH_FRONT = 'PATH_FRONT',
  PATH_FRONT_CONFIRM_EMAIL = 'PATH_FRONT_CONFIRM_EMAIL',
}

export enum GeneralStatus {
  ACTIVATE = 1,
  INACTIVATE = 0,
}