// src/core/constants/translator.enum.ts
export enum TranslatorEnum {
  USER_SUCCESS_CONFIRM = 'auth.USER_SUCCESS_CONFIRM',
  USER_SUCCESS_DELETE = 'user.USER_SUCCESS_DELETE',
  USER_ALREADY_CONFIRM = 'auth.USER_ALREADY_CONFIRM',
  USER_EMAIL_NOT_FOUND = 'auth.USER_EMAIL_NOT_FOUND',
  USER_NEEDS_VERIFY_RESTORE_PWD = 'auth.USER_NEEDS_VERIFY_RESTORE_PWD',
  USER_NEEDS_CHANGE_PWD = 'auth.USER_NEEDS_CHANGE_PWD',
  USER_PASSWORD_CHANGED_SUCCESS = 'auth.USER_PASSWORD_CHANGED_SUCCESS',
  USER_PASSWORD_RESTORED = 'auth.USER_PASSWORD_RESTORED',
  INVALID_CREDENTIALS = 'auth.INVALID_CREDENTIALS',
  TITLE_VALIDATE_EMAIL = 'auth.TITLE_VALIDATE_EMAIL',
  TITLE_RESTORE_PWD = 'auth.TITLE_RESTORE_PWD',
  USER_ALREADY_EXISTS = 'auth.USER_ALREADY_EXISTS',
  EMAIL_SUBJECT_WELCOME = 'auth.EMAIL_SUBJECT_WELCOME',
  SESSION_NOT_FOUND = 'auth.SESSION_NOT_FOUND',
  SESSION_DELETED = 'auth.SESSION_DELETED',
  SESSION_UPDATED = 'auth.SESSION_UPDATED',
  UNAUTHORIZED = 'auth.UNAUTHORIZED',
  UNEXPECTED_SESSION = 'auth.UNEXPECTED_SESSION',
  ROLE_NOT_FOUND = 'user.ROLE_NOT_FOUND',
  ROLE_ALREADY_EXISTS = 'user.ROLE_ALREADY_EXISTS',
  // Validator classes
  VALIDATOR_STRING_FIRST_NAME = 'user.VALIDATOR_STRING_FIRST_NAME',
  VALIDATOR_STRING_LAST_NAME = 'user.VALIDATOR_STRING_LAST_NAME',
  VALIDATOR_STRING_NUMID = 'user.VALIDATOR_STRING_NUMID',
  VALIDATOR_STRING_EMAIL = 'user.VALIDATOR_STRING_EMAIL',
  VALIDATOR_STRING_PHONE = 'user.VALIDATOR_STRING_PHONE',
  VALIDATOR_STRING_ADDRESS = 'user.VALIDATOR_STRING_ADDRESS',
  VALIDATOR_STRING_PASSWORD = 'user.VALIDATOR_STRING_PASSWORD',
  VALIDATOR_STRING_BIRTHDATE = 'user.VALIDATOR_STRING_BIRTHDATE',
  VALIDATOR_STRING_USERNAME = 'user.VALIDATOR_STRING_USERNAME',
  //
  PERMIT_RESOURCE_NOT_FOUND = 'resource.PERMIT_RESOURCE_NOT_FOUND',
  RESOURCE_NOT_FOUND = 'resource.RESOURCE_NOT_FOUND',
  PERMISSION_NOT_FOUND = 'resource.PERMISSION_NOT_FOUND',
  PERMIT_RESOURCE_ALREADY_EXISTS = 'resource.PERMIT_RESOURCE_ALREADY_EXISTS',
  PERMIT_RESOURCE_ASSIGN_SUCCESSFULLY = 'resource.PERMIT_RESOURCE_ASSIGN_SUCCESSFULLY',
  MAILER_ENABLE = 'mailer.MAILER_ENABLE',
  MAILER_DISABLE = 'mailer.MAILER_DISABLE',
  SETTING_INCORRECT_VALUE = 'setting.SETTING_INCORRECT_VALUE',
  INVALID_TOKEN = 'system.INVALID_TOKEN',
  // Surveys
  SURVEY_NOT_FOUND = 'survey.SURVEY_NOT_FOUND',
  SURVEY_QUESTION_NOT_FOUND = 'survey.SURVEY_QUESTION_NOT_FOUND',
  SURVEY_SUCCESS_DELETE = 'survey.SURVEY_SUCCESS_DELETE',
  SURVEY_QUESTION_SUCCESS_DELETE = 'survey.SURVEY_QUESTION_SUCCESS_DELETE',
  SURVEY_ANSWER_NOT_FOUND = 'survey.SURVEY_ANSWER_NOT_FOUND',
  SURVEY_ANSWER_SUCCESS_DELETE = 'survey.SURVEY_ANSWER_SUCCESS_DELETE',
  SURVEY_USER_ANSWER_NOT_FOUND = 'survey.SURVEY_USER_ANSWER_NOT_FOUND',
  SURVEY_USER_ANSWER_SUCCESS_DELETE = 'survey.SURVEY_USER_ANSWER_SUCCESS_DELETE',
}